+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - New `/batches` API for running multi-statement SQL. This replaces `/query-result` API, and is written in a more RESTful approach, removing the need to extend SQLPad timeouts. See API docs for more info.
 
+- Adds `allowedDomains` config item and deprecates `whitelistedDomains` to be removed in v6.
+
 #### Breaking
 
 - `/download-results/` API has been removed in favor of `/statement-results/`, which is similar but based on `statementId` instead of `cacheKey`. See API docs for more info.